Use simpler `hatch_build.py`-based approach
